finished adding hooks to complex components and a little visual fix on TagInput component
