+++ src/components/Spinner/index.tsx
@@ -1,6 +1,7 @@
 import spinnerStyles from './Spinner.module.css';
 
 export const Spinner = ({ size, submitted = false}: any) => {
+	
 	const getSpinnerSize = (size:string) => {
 		const propToCheck = `spinner-${size}`;
 		return spinnerStyles?.[propToCheck] ? spinnerStyles[propToCheck] : '';
